# apps/server/public/api/openapi3.json
{"openapi":"3.0.0","paths":{"/account":{"post":{"operationId":"register","summary":"Register account","description":"","tags":["account"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterAccount"}}}},"responses":{"200":{"description":"Account was successfully registered in system.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/account/change-password":{"post":{"operationId":"change-password","summary":"","description":"Changes the user's password","tags":["account"],"parameters":[],"responses":{"201":{"description":""}}}},"/account/change-email":{"post":{"operationId":"change-email","summary":"","description":"Changes the user's email","tags":["account"],"parameters":[],"responses":{"201":{"description":""}}}},"/account/change-username":{"post":{"operationId":"change-username","summary":"","description":"Changes the user's username","tags":["account"],"parameters":[],"responses":{"201":{"description":""}}}},"/account/delete-domain":{"post":{"operationId":"delete-domain","summary":"","description":"Deletes the user's domain","tags":["account"],"parameters":[],"responses":{"201":{"description":""}}}},"/account/recovery/forgot-password":{"post":{"operationId":"forgot-password","summary":"Request password reset","description":"Sends a password reset email","tags":["account-recovery"],"parameters":[],"responses":{"201":{"description":""}}}},"/account/recovery/reset-password":{"post":{"operationId":"reset-password","summary":"Reset password","description":"Resets the user's password","tags":["account-recovery"],"parameters":[],"responses":{"201":{"description":""}}}},"/account/verification/verify-email":{"post":{"operationId":"verify-email","summary":"Verify Email","description":"Verifies the user’s email","tags":["account"],"parameters":[{"name":"verification_code","required":true,"in":"query","example":"verification_code","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/account/verification/resend-verification-email":{"post":{"operationId":"resend-verification-email","summary":"Re-send verification email","description":"Resends the verification email","tags":["account"],"parameters":[{"name":"email","required":true,"in":"query","description":"Email of created account to which verification email should be sent","example":"Ara_Watsica@hotmail.com","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/authenticate":{"post":{"operationId":"authenticate","summary":"","description":"Logs the user in","tags":["authentication"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Auhenticate"}}}},"responses":{"201":{"description":"The user has been successfully authenticated and session was created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}},"404":{"description":"The user could not be found."}},"security":[{"basic":[]}]},"get":{"operationId":"whoami","summary":"","description":"Returns the current user","tags":["authentication"],"parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/authenticate/refresh-token":{"post":{"operationId":"refresh-token","summary":"","description":"Use a refresh token to extend a session and generate another access token","tags":["authentication"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Auhenticate"}}}},"responses":{"201":{"description":""}}}},"/health":{"get":{"operationId":"healthcheck","summary":"Health check endpoint","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}}}},"/product":{"post":{"operationId":"create-product","summary":"Create product","tags":["product"],"parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"list-products","summary":"List products","tags":["product"],"parameters":[],"responses":{"200":{"description":""}}}},"/product/{id}":{"get":{"operationId":"get-product","summary":"Get product","tags":["product"],"parameters":[],"responses":{"200":{"description":""}}},"patch":{"operationId":"ProductController_updateProduct","parameters":[],"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductController_deleteProduct","parameters":[],"responses":{"200":{"description":""}}}},"/cart":{"get":{"operationId":"CartController_getCart","parameters":[],"responses":{"200":{"description":""}}}},"/cart/{id}":{"put":{"operationId":"CartController_updateCart","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"CartController_deleteCart","parameters":[],"responses":{"201":{"description":""}}}},"/cart/{id}/checkout":{"post":{"operationId":"CartController_checkoutCart","parameters":[],"responses":{"201":{"description":""}}},"delete":{"operationId":"CartController_cancelCheckoutCart","parameters":[],"responses":{"200":{"description":""}}}},"/payment/{id}/pay":{"post":{"operationId":"PaymentController_payPayment","parameters":[],"responses":{"201":{"description":""}}}},"/payment":{"get":{"operationId":"PaymentController_listPayments","parameters":[],"responses":{"200":{"description":""}}}},"/payment/{id}":{"get":{"operationId":"PaymentController_getPayment","parameters":[],"responses":{"200":{"description":""}}}},"/stripe/webhook":{"post":{"operationId":"StripeWebhookController_handleWebhook","parameters":[{"name":"stripe-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}}},"info":{"title":"methylophenidate","description":"Methylophenidate is a boilerplate for Nest.js applications with batteries included.","version":"1.0","contact":{}},"tags":[{"name":"api","description":""}],"servers":[],"components":{"schemas":{"AuthenticationResponse":{"type":"object","properties":{"id":{"type":"string","description":"ID of authenticated domain."},"mfa":{"type":"boolean","description":"Indicates if domain uses MFA. In case of true, additional steps are supposed to be made to activate access token."},"accessToken":{"type":"string","description":"Access token."},"refreshToken":{"type":"string","description":"Refresh token."}},"required":["id","mfa","accessToken","refreshToken"]},"RegisterAccount":{"type":"object","properties":{"email":{"type":"string","description":"The domain's email address","example":"Ara_Watsica@hotmail.com","examples":["Raul_Marks@gmail.com","Buddy.Schulist@gmail.com","Anibal_Pouros@hotmail.com"]},"password":{"type":"string","description":"The domain's password","example":"SuperSecretPassword123!","examples":["LaW1KVuo69FG5Pf","9srQfkporSDhU76","74ETrfy3OQCdUzs"]},"username":{"type":"string","description":"The domain's username","example":"Sigurd.Mohr56","examples":["Raul_Marks","Buddy.Schulist","Anibal_Pouros"]}},"required":["email","password","username"],"description":"asdasd"},"Account":{"type":"object","properties":{"id":{"type":"string","description":"The domain's unique identifier","example":"cjld2cjxh0000qzrmn831i7rn"},"email":{"type":"string","description":"The domain's email address","example":"Ara_Watsica@hotmail.com","examples":["Raul_Marks@gmail.com","Buddy.Schulist@gmail.com","Anibal_Pouros@hotmail.com"]},"emailVerified":{"type":"boolean","description":"Indicates whether the email associated with a user domain has been verified","example":false},"username":{"type":"string","description":"The domain's username","example":"Sigurd.Mohr56","examples":["Raul_Marks","Buddy.Schulist","Anibal_Pouros"]}},"required":["id","email","emailVerified","username"],"description":"asdasd"},"Auhenticate":{"type":"object","properties":{"username":{"type":"string","description":"The domain's username","example":"Sigurd.Mohr56","examples":["Arnulfo_Fisher","Ella.Schumm","Stefan78"]},"password":{"type":"string","description":"The domain's password","example":"SuperSecretPassword123!","examples":["qjC7BP4z25WKeaw","OVhG3k4wT2rrVnM","W1bzTeoy5KY18vT"]}},"required":["username","password"],"description":"asdasd"}}}}